backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
prometheus_client
pytest
httpx
python-multipart